implement auth middleware

+++ backend/src/controllers/notesController.ts
@@ -2,11 +2,13 @@ import { RequestHandler } from "express";
 import NoteModel from "../models/note";
 import mongoose from "mongoose";
 import createHttpError from "http-errors";
+import { assertIsDefined } from "../util/assertIsDefined";
 
 export const getNotes: RequestHandler = async (req, res, next) => {
+    const authenticatedUserId = req.session.userId;
     try {
-        // throw createHttpError(401);
-        const notes = await NoteModel.find().exec();
+        assertIsDefined(authenticatedUserId); // to make sure user id is valid with generic type check
+        const notes = await NoteModel.find({userId: authenticatedUserId}).exec();
         res.status(200).json(notes);
     } catch (error) {
         next(error);
@@ -15,7 +17,9 @@ export const getNotes: RequestHandler = async (req, res, next) => {
 
 export const getNote: RequestHandler = async (req, res, next) => {
     const noteId = req.params.noteId;
+    const authenticatedUserId = req.session.userId;
     try {
+        assertIsDefined(authenticatedUserId);
         if (!mongoose.isValidObjectId(noteId)) {
             throw createHttpError(400, "Invalid note ID");
         }
@@ -23,6 +27,10 @@ export const getNote: RequestHandler = async (req, res, next) => {
 
         if (!note) {
             throw createHttpError(404, "Note not found!");
+        }
+
+        if (!note.userId.equals(authenticatedUserId)) {
+            throw createHttpError(401, "You cannot access this note");
         }
         res.status(200).json(note);
     } catch (error) {
@@ -37,11 +45,14 @@ interface CreateNoteBody {
 export const createNotes: RequestHandler<unknown, unknown, CreateNoteBody, unknown> = async (req, res, next) => {
     const title = req.body.title;
     const text = req.body.text;
+    const authenticatedUserId = req.session.userId;
     try {
+        assertIsDefined(authenticatedUserId);
         if (!title) {
             throw createHttpError(400, "Note must have a title")
         }
         const newNote = await NoteModel.create({
+            userId: authenticatedUserId,
             title: title,
             text: text,
         });
@@ -64,8 +75,10 @@ export const updateNote: RequestHandler<UpdateNoteParams, unknown, UpdateNoteBod
     const noteId = req.params.noteId;
     const newTitle = req.body.title;
     const newText = req.body.text;
+    const authenticatedUserId = req.session.userId;
 
     try {
+        assertIsDefined(authenticatedUserId);
         if (!mongoose.isValidObjectId(noteId)) {
             throw createHttpError(400, "Invalid note ID");
         }
@@ -79,6 +92,10 @@ export const updateNote: RequestHandler<UpdateNoteParams, unknown, UpdateNoteBod
             throw createHttpError(404, "Note not found");
         }
 
+        if (!note.userId.equals(authenticatedUserId)) {
+            throw createHttpError(401, "You cannot access this note")
+        }
+
         note.title = newTitle;
         note.text = newText;
         const updatedNote = await note.save();
@@ -90,8 +107,10 @@ export const updateNote: RequestHandler<UpdateNoteParams, unknown, UpdateNoteBod
 
 export const deleteNote: RequestHandler = async (req, res, next) => {
     const noteId = req.params.noteId;
+    const authenticatedUserId = req.session.userId;
 
     try {
+        assertIsDefined(authenticatedUserId);
         if (!mongoose.isValidObjectId(noteId)) {
             throw createHttpError(400, "Invalid note ID");
         }
@@ -101,6 +120,10 @@ export const deleteNote: RequestHandler = async (req, res, next) => {
             throw createHttpError(404, "Note not found");
         }
 
+        if (!note.userId.equals(authenticatedUserId)) {
+            throw createHttpError(401, "You cannot access this note")
+        }
+        
         await note.deleteOne();
         res.sendStatus(204);
     } catch (error) {
